Update Notification example to latest template
