+++ Examples/NotificationExample/plcnext.proj
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <ProjectSettings xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema" xmlns="http://www.phoenixcontact.com/schema/cliproject">
-  <Target>AXCF2152,2020.6.1 (20.6.1.30859)</Target>
+  <Target>AXCF2152,2021.0.0 LTS (21.0.0.35466)</Target>
   <Id>d2650cb2-0629-4666-91e9-f94d07a80772</Id>
   <Version>1.1</Version>
 </ProjectSettings>
